Rewrite the request interceptor.
It seems that there are request in the form
"blob://foo.zim/<uuid>".
I don't know where it come from, we cannot handle it and even Qt is parsing
it wrongly (scheme and host are empty, path is the full url request).

By using a string to handle the url, it is simpler to change "http://" to
"zim://".