Set the tab's icon to the icon of the zim file.
Properly set the icon of the tab when the favicon of the KiwixWebView
is updated.

By default, the icon of QWebEngineView is updated from the favicon set
in the html. But it sometime it is not set.

So we need to load the icon from the zim file, so we have to react
to urlChanged signal, to load the icon when the url's host (the zim file)
changes.