Use real url for icon in the contentManager instead of data url.
The update of book info are made somehow sync. (They are async, but
somehow, the view update hangout).

Getting the favicon content is pretty long and it seems that the view
(and getting remote content) take a long time.

By using a real url, QWebViewEngine handle everything nicely and it speed
up everything.