Use the real zimId instead of "zimId+.zim".
The real zimId is `xxxx` (stored in the zim itself), not `xxxx.zim` (we
use as host url).

Store the host as an id means that we always need to convert from `xxxx` to
`xxxx.zim` (and the contrary). It is prone to mistake when we forget to
convert to the correct format.

We must use `xxxx.zim` only in the url as is it use to identify what is
asked in the urlschemehandler. Everywhere else, we only need the real
zimId.