Add a filtering by language.
The current list of available language is a fixed list.
It is generated by the union of languages we have (on library.kiwix.org)
and languages Qt know of. So some languages for which we have content
are not available.

For now, the filtering is made only for remote catalog.
This probably have to be changed, but for now it is better to
display all the content the user has.