Keep the mp_stackWidget in sync with the tabBar.
We need to remove the child widget from the stackWidget before we remove
the tab. Else the onCurrentChanged will try to get the widget of the new
tab and if the tab and the stack are not in sync, we may get the wrong
widget.

As `closeTab` remove the widget from the stack we don't need to remove
it again in listen of `CloseTabAction`.