Fix handleMetaRequest.
The main fix is about to return after we reply the request and not do a
`request->fail` just after.
The try/catch is just here be safety.

It seems that depending of the version of Qt, doing a `reply` followed by
a `fail`, the request is replied (as with Qt packaged in fedora) or
failed (as with Qt packaged in ubuntu).

This should fix #239