Add a default zoom field in the settings
The SettingsManager class has a new member m_zoomFactor with a getter and
a setter. Its value is set at launch and saved with the QSettings.

There is a new field in the settings page where the zoom factor can be
changed, it is multiplied by 100 to be displayed as a percentage.

Each time a file is opened, the zoomfactor of the view is set with the
member m_zoomFactor of the SettingsManager class.