Fix pagination by setting pageLength properly
The pageLength parameter of kiwix::SearchRenderer is not set properly
which causes it to take an arbitrary large value breaking pagination.

The solution is to use the same convention as kiwix-serve. Try to read
the pageLength from query if found, else use a default value of 25. Use
this pageLength to find end. Use renderer.setPageLength() method to set
the pageLength properly.