Set the fullscreen border after the windows is shown.
`windowHandle` return a valid value only if the widget has been shown.
So we must use it *after* it is shown, not at the end of its construction.

This subtle change also avoid a early return in the MainWindow constructor
and so, we correctly configure the mainWindow on Windows.
The last configuration is setting the content manager of the content manager
side. If this is skip, kiwix-desktop crash as soon as we click on "All files"
as we use a uninitialized pointer.