Don't monitor files while in All files view.
When downloading a file, with monitor directory set same as download directory, the changes in fs may send a double booksChanged signal.
The first is due to adding/removing books while the second is the explicit booksChanged.
This change stops monitoring while in All files view.
To save the edge case when the user themselves adds books in monitor directory (while Kiwix is in All files view), a connection is added to Local Files button - to call loadMonitorDir().