Move the server management in kiwixapp.
`LocalKiwixServer` is "just" a UI to start/configure or stop the server.

One important change is that instead of having one server we configure and
start/stop when needed, we are now creating a new server each time
we want to start the server.

It doesn't change the behavior, as libkiwix is destroying/recreating for
us, but it will help us to use the new API which use a
`Server::Configuration`.