Bugfix: completed downloads are removed
The commit "Moved download deregistration to ContentManager" introduced
a bug - completed downloads were not properly deregistered.

Note that a simpler fix by adding a single `m_downloads.remove(bookId);`
line in `ContentManager::downloadCompleted()` didn't work (which means
that that function is not always called upon download completion). I am
not going to investigate that issue now, hoping that it will be
automatically resolved once the redesign of download management is
finished.