!!! Disabled monitoring of the download dir
Monitoring of the download dir interferes with the download lifecycle
management. Here are some of the observed effects:

- As soon as the ZIM file is created in the download directory it
  is added to the library and overwrites the entry created by download
  management logic, thus resetting the download id associated with the book.

- As a result `ContentManager::downloadCompleted()` is never called;
  `ContentManager::downloadCancelled()` is called instead (however, for
  some reason the user-observable effect is as if the download completion
  is normally detected and handled). Looks like this was the main
  culprit behind the false bug report kiwix/libkiwix#1049.

- The icon of the book added a time when the file was not yet a valid
  ZIM file stays invalid.

I am temporarily disabling this functionality so that the work on
download related code is carried out in a clean room environment.
Once it's complete monitoring of the download directory must be
re-enabled but in a way that doesn't confuse the logic of the
download state machine.