Reporting of issues with the download directory
Before this change a non-existent download directory would result in
"The system doesn't have enough storage available" error. Trying to save
files to non-writable directory resulted in the download seemingly being
stuck (but this is just a special case of not properly handling the download
error status returned by Aria).