Handling of invalid ZIM files
When trying to display resources from an invalid ZIM file that fail to
be retrieved due to problems with the ZIM data, now an HTML error
page is returned instead. Note that the resource being accessed
may in fact be an image, a CSS or a javascript file and the user may not see
the error (though the appearance of the webpage trying to load and use
that resource will be affected to various degrees).