Fixed CI build
In CI build, dependencies are put in the build install location rather than
in the system libs directory, hence that path must be included in the
search for shared libraries.

The problem seems to have been triggered by recent change in the
libkiwix.pc pkg-config file which resulted in the version of some
of libkiwix dependencies (recorded in libkiwix.so) being bumped up.
As a result, previous/older versions of those dependencies embedded in
the CI build docker image
(ghcr.io/kiwix/kiwix-build_ci_focal:2024-06-03) no longer matched the
requirements of libkiwix.so, whereas the up-to-date versions of those
dependencies (that are downloaded during each build procedure) were not
used because they were not in the search path.