Updating App Icons
New version of the App Icon: Instead of the black Kiwix bird on transparent (or white)
background, this new version is a lightly-rounded square with a white to grey gradient
that serves as a background for a black Kiwix bird on top.
The main goal of this change is for the icon to be clearly visible on any background;
given OS now support dark/light and custom modes.
I looks quite similar to the iOSmacOS and the Android icons.

In addition, those are now properly resized to all the existing sizes plus a new 512 size.

A new custom Icon for ZIM files is also included and declared for all sizes.
This seems not to be in use or broken but at least the new file is there.
It's a (macOS style) blank document (corner-folded paper) with the black bird on top
with the ZIM label at the bottom