[HACK] Handling of empty result after elision
The bug itself happens due to initialization&update of MultiZimButton (via
established signal/slot connections) at a time when the main application
window hasn't yet been shown and the dimensions of various widgets are
bogus. The end result is that a too narrow menu leads to the elided text
string being empty which was outside of the assumptions of the
getElidedText() function.