Restore tabs only after disabled sandbox is approved
Before this fix tabs could be restored before the user approved
disabling the QWebEngine sandbox (when the app is launched from a remote
location/network folder). If disabling the web browser sandbox in
kiwix-desktop has any real risk for the user then we shouldn't expose
them to it before notifying about implications and receiving their
confirmation to proceed.