Fixed double rendering of suggestions
PR #1414 resulted in suggestions being rendered twice

1. Default painting that was intended to be overriden by custom rendering
   introduced by #1189 (this is needed to automatically highlight the
   selected item)

2. followed by custom painting with some offset

PR #1414 had created a false impression that the default rendering of
the icon and text was suppressed by setting those fields of
QStyleOptionViewItem to empty values, whereas the text and icon in fact
were obtained from the QModelIndex. The PR should have been tested as
part of the review!

Now the bug is worked around by moving all the displayable data of
SuggestionListModel entries to UserRole, while returning empty values
for DisplayRole and DecorationRole data.